extracted naming.go from remaster_plan.go
